It worksgit add config/schedule.rb !

+++ app/workers/thread_pruning_worker.rb
@@ -3,24 +3,28 @@ class ThreadPruningWorker
 
   #removes threads and posts which are no longer active
   def prune_posts(board)
-    now = Time.now
-    puts "Pruning Posts at #{now}"
+
     tracked_threads = {}
     Post.select(:op).distinct.where(board: board).each{|e| tracked_threads.merge!({e.op => false})}
     current_threads = Fourchan::Kit::Board.new(board).all_threads
     current_threads.each{ |t| tracked_threads[t.no.to_s] = true }
     tracked_threads.each_key do |k|
       #binding.pry
+      #p k
       next if tracked_threads[k] == true
+
       old_posts = Post.where(op: k)
       old_posts.each do |o|
+        #p o
         old_counter = PostCounter.find_by(text_hash: o.text_hash)
+        next if old_counter.nil?
         old_counter[:occurrences] -= 1
         old_counter[:occurrences] > 0 ? old_counter.save : old_counter.destroy
       end
       old_posts.destroy_all
     end
-    puts "Finished pruning post at #{Time.now}\nDuration: #{Time.now - now}\n\nThere are currently #{Post.count} tracked posts"
+    num_pruned = 0
+    tracked_threads.each_value{|v| v ? next : num_pruned += 1}
 
   end
   #handle_asynchronously :purge_old_posts
